Rename to gliderflightOG1
