+++ .gitignore
@@ -173,3 +173,7 @@ cython_debug/
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 .idea/
+
+# Claude Code files
+CLAUDE.md
+.claude/*
