+++ docs/source/conf.py
@@ -17,7 +17,7 @@ year = datetime.datetime.now(tz=datetime.timezone.utc).date().year
 # print(sys.path)
 
 # General information about the project.
-project = "flightmodels"
+project = "gliderflightOG1"
 author = "Eleanor Frajka-Williams, FirstName LastName"
 copyright = f"{year}, {author}"
 release = "v0.0.0"
