+++ docs/source/gliderflightOG1.rst
@@ -1,39 +1,39 @@
-:mod:`flightmodels API`
+:mod:`gliderflightOG1 API`
 ---------------------------
 
 readers
 =======
-.. automodule:: flightmodels.readers
+.. automodule:: gliderflightOG1.readers
    :members:
    :undoc-members:
 
 flight
 ======
-.. automodule:: flightmodels.flight
+.. automodule:: gliderflightOG1.flight
    :members:
    :undoc-members:
 
 plotters
 ========
-.. automodule:: flightmodels.plotters
+.. automodule:: gliderflightOG1.plotters
    :members:
    :undoc-members:
 
 writers
 =======
-.. automodule:: flightmodels.writers
+.. automodule:: gliderflightOG1.writers
    :members:
    :undoc-members:
 
 tools
 =====
-.. automodule:: flightmodels.tools
+.. automodule:: gliderflightOG1.tools
    :members:
    :undoc-members:
 
 utilities
 =========
-.. automodule:: flightmodels.utilities
+.. automodule:: gliderflightOG1.utilities
    :members:
    :undoc-members:
