+++ docs/source/setup.md
@@ -7,14 +7,14 @@ To install this package, currently not released to pypi or conda-forge, you need
 Clone the repository and install it locally using `pip install -e .`
 
 ```bash
-git clone https://github.com/eleanorfrajka/flightmodels.git
-cd flightmodels
+git clone https://github.com/eleanorfrajka/gliderflightOG1.git
+cd gliderflightOG1
 pip install -r requirements-dev.txt
 pip install -e .
 ```
 
 An alternative for the first step (`clone`) would be, in your **browser**:
-1. Navigate to [https://github.com/eleanorfrajka/flightmodels](https://github.com/eleanorfrajka/flightmodels)
+1. Navigate to [https://github.com/eleanorfrajka/gliderflightOG1](https://github.com/eleanorfrajka/gliderflightOG1)
 2. Click the green `<> Code` button.
 3. Choose **Open with GitHub Desktop**.
