+++ environment.yml
@@ -1,0 +1,10 @@
+name: gliderflightOG1
+channels:
+  - conda-forge
+dependencies:
+  - python>=3.9
+  - pip
+  - pandoc
+  - pip:
+      - -r requirements.txt
+      - -r requirements-dev.txt
