+++ gliderflightOG1/flightmodels.mplstyle
