+++ gliderflightOG1/utilities.py
@@ -6,8 +6,8 @@ from urllib.parse import urlparse
 import xarray as xr
 import requests
 
-from flightmodels import logger
-from flightmodels.logger import log_debug, log_error, log_info
+from gliderflightOG1 import logger
+from gliderflightOG1.logger import log_debug, log_error, log_info
 
 log = logger.log
